Code refactoring, additional input checks and unit tests
Replaced most '/n' with std::endl as speed is not important in those contexts, but showing the messages even if the program crashes afterwards is
Added maximum number of shown errors to FragmentDistributionStats::UpdateRefSeqBias
Check if reference bias from file is positive
Sorted utilities alphabetical
Check if more than 64 alleles
Replaced all local solutions with utilities::SafePercent, which returns 50 for zero denominator
For Reference::ReadVariants maximum errors is now set in a constant and when reached it will be printed
Replaced local solutions with TransformDistanceToStartOfErrorRegion to do the +9/10 transformation to start of error region
Constants in classes are now named after google style kMyImportantClassVariable and are static const if other classes need to use them
Added IsGC and IsN to utilities and replaced local solutions
Added unit tests for Sign, IsGC and IsN from utilities
Replaced ModifiedString with own typedefs in utilities
Shifted ReferenceTest to the same level as SeqQualityStatsTest, so we don't get tons of subsequent errors blocking the Reference informations
Added a Reference::GCContentAbsolut without n_count parameter
Removed utilities.cpp as utilities is header only anyways, because we have only short inline functions and template classes
Typedef for integers based on usecase to unify, show and option to quickly change
Moved ReverseComplementor from Reference to utilties
Check if ambiguous characters (N's) are in variant description
Checking for reference sequences and number of reference sequences larger than uint32
Moved Complement from Reference into utilities

+++ reseq/QualityStatsTest.h
@@ -20,10 +20,10 @@ namespace reseq{
 
 		template<typename T> static void TestSizeAtIndex_0_1_98_99(
 				T& vector,
-				uint64_t size0,
-				uint64_t size1,
-				uint64_t size98,
-				uint64_t size99,
+				size_t size0,
+				size_t size1,
+				size_t size98,
+				size_t size99,
 				const char *context,
 				const char *err_msg
 				){
@@ -35,10 +35,10 @@ namespace reseq{
 
 		template<typename T> static void TestValueAtIndex_0_1_98_99(
 				T& vector,
-				uint64_t val0,
-				uint64_t val1,
-				uint64_t val98,
-				uint64_t val99,
+				size_t val0,
+				size_t val1,
+				size_t val98,
+				size_t val99,
 				const char *context,
 				const char *err_msg
 				){
